Refactor device controller #985
