+++ src/AzureIoTHub.Portal/Server/Services/DevicePropertyService.cs
@@ -1,0 +1,145 @@
+// Copyright (c) CGI France. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace AzureIoTHub.Portal.Server.Services
+{
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Threading.Tasks;
+    using Azure;
+    using Entities;
+    using Exceptions;
+    using Factories;
+    using Helpers;
+    using Models.v10;
+    using Newtonsoft.Json;
+    using Newtonsoft.Json.Linq;
+
+    public class DevicePropertyService : IDevicePropertyService
+    {
+        private readonly IDeviceService devicesService;
+        private readonly ITableClientFactory tableClientFactory;
+
+        public DevicePropertyService(IDeviceService devicesService, ITableClientFactory tableClientFactory)
+        {
+            this.devicesService = devicesService;
+            this.tableClientFactory = tableClientFactory;
+        }
+
+
+        public async Task<IEnumerable<DevicePropertyValue>> GetProperties(string deviceId)
+        {
+            var device = await this.devicesService.GetDeviceTwin(deviceId);
+
+            if (device == null)
+            {
+                throw new ResourceNotFoundException($"Unable to find the device {deviceId}");
+            }
+
+            var modelId = DeviceHelper.RetrieveTagValue(device, nameof(DeviceDetails.ModelId));
+
+            if (string.IsNullOrEmpty(modelId))
+            {
+                throw new ResourceNotFoundException($"Device {deviceId} has no modelId tag value");
+            }
+
+            AsyncPageable<DeviceModelProperty> items;
+
+            try
+            {
+                items = this.tableClientFactory
+                    .GetDeviceTemplateProperties()
+                    .QueryAsync<DeviceModelProperty>($"PartitionKey eq '{modelId}'");
+            }
+            catch (RequestFailedException e)
+            {
+                throw new InternalServerErrorException($"Unable to get templates properties fro device with id {deviceId}: {e.Message}", e);
+            }
+
+            var result = new List<DevicePropertyValue>();
+            JObject desiredPropertiesAsJson;
+            JObject reportedPropertiesAsJson;
+
+            try
+            {
+                desiredPropertiesAsJson = JObject.Parse(device.Properties.Desired.ToJson());
+            }
+            catch (JsonReaderException e)
+            {
+                throw new InternalServerErrorException($"Unable to read desired properties for device with id {deviceId}", e);
+            }
+
+            try
+            {
+                reportedPropertiesAsJson = JObject.Parse(device.Properties.Reported.ToJson());
+            }
+            catch (JsonReaderException e)
+            {
+                throw new InternalServerErrorException($"Unable to read reported properties for device with id {deviceId}", e);
+            }
+
+            await foreach (var item in items)
+            {
+                var value = item.IsWritable ? desiredPropertiesAsJson.SelectToken(item.Name)?.Value<string>() :
+                        reportedPropertiesAsJson.SelectToken(item.Name)?.Value<string>();
+
+                result.Add(new DevicePropertyValue
+                {
+                    DisplayName = item.DisplayName,
+                    IsWritable = item.IsWritable,
+                    Name = item.Name,
+                    PropertyType = item.PropertyType,
+                    Value = value
+                });
+            }
+
+            return result;
+        }
+
+        public async Task SetProperties(string deviceId, IEnumerable<DevicePropertyValue> devicePropertyValues)
+        {
+            var device = await this.devicesService.GetDeviceTwin(deviceId);
+
+            if (device == null)
+            {
+                throw new ResourceNotFoundException($"Unable to find the device {deviceId}");
+            }
+
+            var modelId = DeviceHelper.RetrieveTagValue(device, nameof(DeviceDetails.ModelId));
+
+            if (string.IsNullOrEmpty(modelId))
+            {
+                throw new ResourceNotFoundException($"Device {deviceId} has no modelId tag value");
+            }
+
+            AsyncPageable<DeviceModelProperty> items;
+
+            try
+            {
+                items = this.tableClientFactory
+                    .GetDeviceTemplateProperties()
+                    .QueryAsync<DeviceModelProperty>($"PartitionKey eq '{modelId}'");
+            }
+            catch (RequestFailedException e)
+            {
+                throw new InternalServerErrorException($"Unable to get templates properties fro device with id {deviceId}: {e.Message}", e);
+            }
+
+            var desiredProperties = new Dictionary<string, object>();
+
+            await foreach (var item in items)
+            {
+                if (!item.IsWritable)
+                {
+                    continue;
+                }
+
+                _ = desiredProperties.TryAdd(item.Name, devicePropertyValues.FirstOrDefault(x => x.Name == item.Name)?.Value);
+            }
+
+            device.Properties.Desired = DeviceHelper.PropertiesWithDotNotationToTwinCollection(desiredProperties);
+
+            _ = await this.devicesService.UpdateDeviceTwin(device);
+        }
+    }
+}
