+++ src/AzureIoTHub.Portal/Server/Services/IDevicePropertyService.cs
@@ -1,0 +1,16 @@
+// Copyright (c) CGI France. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace AzureIoTHub.Portal.Server.Services
+{
+    using Models.v10;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
+
+    public interface IDevicePropertyService
+    {
+        Task<IEnumerable<DevicePropertyValue>> GetProperties(string deviceId);
+
+        Task SetProperties(string deviceId, IEnumerable<DevicePropertyValue> devicePropertyValues);
+    }
+}
